requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.13.2
pydantic==2.11.7
python-dotenv==1.1.1
yfinance==0.2.65
requests==2.32.4
schedule==1.2.2
beautifulsoup4==4.13.4
pandas==2.3.1
numpy==2.0.2
